general cleanup on modules/webgl
remove old code
cleanup types
remove test components, pages and state management
