+++ apps/playground/src/modules/webgl/components/ConfigTable/icons.ts
@@ -1,8 +1,8 @@
-import { TControllerType } from "modules/webgl/controllers/InputController/types";
-import { Icon, Keyboard, Mouse, DeviceGamepad2 } from "tabler-icons-react";
+import { TControllerType } from "modules/webgl/libs/types";
+import { Icon, Keyboard, Mouse } from "tabler-icons-react";
 
 export const deviceIcons: { [key in TControllerType]: Icon } = {
   keyboard: Keyboard,
   mouse: Mouse,
-  gamepad: DeviceGamepad2
+  // gamepad: DeviceGamepad2
 };
