+++ apps/playground/src/modules/webgl/components/ConfigTable/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { createStyles, Table, Text } from "@mantine/core";
 
 
-import type { TConfiguration } from "modules/webgl/controllers/InputController/types";
+import type { TConfiguration } from "modules/webgl/libs/types";
 import {  getMaxInputsCount } from "./helpers";
 import { AxisRows } from "./AxisRows";
 
@@ -17,7 +17,7 @@ const useStyles = createStyles((theme) => ({
   }
 }));
 
-export const ConfigTable: React.FC<{ config: TConfiguration }> = ({
+export const ConfigTable: React.FC<{ config: TConfiguration}> = ({
   config
 }) => {
   const { classes } = useStyles();
