+++ apps/playground/src/modules/webgl/config/index.ts
@@ -1,2 +1,1 @@
-export { CONTROLS } from './inputs';
 export { config } from './config.inputs';
