--- apps/playground/src/modules/webgl/config/inputs.ts
@@ -1,14 +1,0 @@
-import { InputConfig } from "../types";
-
-export const CONTROLS: InputConfig = {
-  fwd: ['KeyW', 'ArrowUp'],
-  left: ['KeyA', 'ArrowLeft'],
-  back: ['KeyS', 'ArrowDown'],
-  right: ['KeyD', 'ArrowRight'],
-  up: ['KeyR'],
-  down: ['KeyF'],
-  boost: ['ShiftLeft'],
-  break: ['Space'],
-  rollLeft: ['KeyQ'],
-  rollRight: ['KeyE'],
-};
